refactoring

+++ src/main/java/org/example/petstore/repository/ProductRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-    Optional<List<Product>> findByCategory_CategoryId(Integer categoryId);
+public interface ProductRepository extends JpaRepository<Product, Long> {
+    Optional<List<Product>> findByCategory_CategoryId(Long categoryId);
 }
